Split out ratio of means link calculation from generic imputation engine code
